// agriculture-app/src/stores/enums/Prefetch.ts
export const enum Prefetch {
  UserFarmInfo = "user_farm_info",
  UserAdjustments = "user_adjustments",
  UserEquipment = "user_equipment",
  UserOperations = "user_operations",
  UserCropTypes = "user_crop_types",
  UserOperationTypeCodifiers = "user_operation_type_codifiers",
  UserAgriculturalOperationCodifiers = "user_agricultural_operations_codifiers",
  UserFarmlands = "user_farmlands",
  FetchAllIndicators = "fetch_all_indicators",
}